declaring refs like this is causing errors
Invariant Violation: addComponentAsRefTo(...): Only a ReactOwner can
have refs. This usually means that you're trying to add a ref to a
component that doesn't have an owner (that is, was not created inside
of another component's `render` method). Try rendering this component
inside of a new top-level component which will hold the ref.

@JedWatson look into
https://facebook.github.io/react/docs/more-about-refs.html#the-ref-callb
ack-attribute